Added Documentation File
I made the documentation file for the project and filled some of it out. I also changed the int cast in the CAN messege to a char cast. horror.py as well is a program I am working on for Sam because he doesn't want to do it himself, I don't blame him.