Allan Variance
Worked on the Allan Variance script a bit, it still needs a lot more work but it currently reads in the data and will make a txt file for easier re-runs. I also worked on the arduino code a bit but the progress there was minimal. Also did a small amount of documentation writing too.