Final push
This is my final push as the project is coming to a close. All of the work in this is on the documentation. If you are apart of a future team and are reading this feel free to reach out to me with any questions about the code, I would be happy to explain it!